Adds Ability Batch Process INSERT Statements
CSVtoSQL was updated to include the ability to point the tool to a
directory containing multiple .csv files for processing, as opposed to
running one at a time. Updated the argument -f to reflect --filepath as opposed to --file as now this flag can be used to define a directory of multiple csv/sql
files for batch processing or file merging.

Added custom error messages and exception handling for certain
situations around the newly added multiple file processing.